First pass of changes to refactor the JIT to merge all modules into a single module before trying to resolve JITted function and global pointers.
This change does a merge immediately after finishing (and optionally caching) a
module.  It also adds a visitor to replace all of the reps in the ModuleDef
and all contained Defs to their corresponding values in the merged module,
removes most of the code that does global variable address mapping, comments
out the resolver and all references to it, and removes binding of modules to the
execution engine.